+++ src/kannushi/_logging.py
@@ -7,6 +7,7 @@ from sys import stderr
 
 class _AnsiColor(str, Enum):
     DEFAULT = '\033[0m'
+    DIM     = '\033[2m'
     RED     = '\033[31m'
     GREEN   = '\033[32m'
     YELLOW  = '\033[33m'
@@ -16,6 +17,7 @@ class _AnsiColor(str, Enum):
 #
 
 _is_color_disabled = False
+_is_verbose        = False
 
 #
 # Interface
@@ -25,9 +27,20 @@ def set_color_disabled(is_color_disabled: bool):
     global _is_color_disabled
     _is_color_disabled = is_color_disabled
 
+def set_verbose(verbose: bool):
+    global _is_verbose
+    _is_verbose = verbose
+
+def is_verbose() -> bool:
+    return _is_verbose
+
 
 def print_success(*args, **kwargs):
     _print_in_color(_AnsiColor.GREEN, *args, **kwargs)
+
+def print_verbose(*args, **kwargs):
+    if _is_verbose:
+        _print_in_color(_AnsiColor.DIM, *args, **kwargs)
 
 def print_warning(*args, **kwargs):
     _print_in_color(_AnsiColor.YELLOW, *args, **dict({'file' : stderr}, **kwargs))